backend/requirements.txt
google-adk==1.14.1
fastapi==0.116.2
uvicorn==0.35.0
python-multipart==0.0.20
python-dotenv==1.1.1
